pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdds"
version = "0.1.0"
description = "A modular framework for discrete diffusion models featuring GDDS (Generalized Discrete Diffusion from Snapshots) and the Semantic-Informed Kernel (SIK)."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Oussama Zekri", email = "oussama.zekri@ensae.fr"}
]
keywords = ["diffusion", "discrete-diffusion", "machine-learning", "deep-learning", "pytorch", "lightning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "lightning==2.5.1",
    "torch==2.7.0",
    "torchmetrics==1.7.1",
    "numpy==1.26.4",
    "einops==0.8.1",
    "fancy-einsum==0.0.3",
    "tqdm==4.67.1",
    "termcolor==3.0.1",
    "fsspec",
    "jinja2==3.1.5",
    "pyyaml==6.0.2",
    "pandas==2.2.1",
    "scikit-learn==1.4.0",
    "timm==1.0.15",
    "tensorboard==2.16.2",
    "wandb==0.19.9",
    "rich==13.9.4",
    # Data / Hugging Face
    "datasets==3.5.0",
    "transformers==4.45.0",
    "tokenizers==0.20.3",
    "huggingface-hub==0.30.2",
    "safetensors==0.5.2",
    "accelerate==1.6.0",
    "sentencepiece==0.2.0",
    "requests>=2.31.0",
    "blobfile==3.0.0",
    "ocifs==1.3.2",
    # Evaluation
    "evaluate==0.4.3",
    "mauve-text==0.4.0",
    "sentence-transformers>=3.0.0",
    "lm-eval @ git+https://github.com/EleutherAI/lm-evaluation-harness",
    # Optional-but-used codepaths
    "pykeops>=2.2",
    "jupyter==1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=25.0.0",
]
flash-attn = [
    "flash-attn>=2.7.0",
]
gpu-kernels = [
    "triton>=3.0.0",
]
mamba = [
    "causal-conv1d==1.4.0",
    "mamba-ssm==1.2.0.post1",
]

[project.scripts]
gdds = "discrete_diffusion.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["discrete_diffusion*"]
namespaces = false
